fix(path): Expand ~ (tilde) in target paths
https://www.shellcheck.net/wiki/SC2088 mentions that tildes are not
expanded within quotes.

This means that an output directory mapped like `a=~/a` will result in
/home/myuser/~/a, where the last "~/a" is a folder, beginning with "~"
and containing a "/", which is confusing.

With this patch, the `a=~/a` binding will result in /home/myuser/a,
which is expected IMO.

A side effect of this, is that an input of `a=~/a` will print out
`/home/myuser/a` in the key mapping preview pane. A later PR will get
rid of this.

A new dependency is added to handle this case, "expanduser".