feat: Try to shorten directory to fit into preview
We use another dependency, tico, to get the "/a/b/c/directory" pattern
of displaying. It has a downside, it only expands the final "/directory"
if we don't end with a "/", so we work around that a bit.

0b79ab2051 replaces ~ with the actual path (like /home/myuser), and this
patch seemingly reverses that behavior. I would argue this is fine
though, since the previous commit makes the internal path state usable
everywhere, even when exporting the run-script. This patch's reversal to
"~" is for display purposes only.